fusor server: refactoring deploy action in to multiple sub-actions
This commit refactors the current 'deploy' action in to 3 sub-actions:
   - Content (Katello/Satellite) (Actions::Fusor::Content::ManageContent)
   - Deploy RHEV (Actions::Fusor::Deployment::Rhev::Deploy)
   - Deploy CloudForms (Actions::Fusor::Deployment::CloudForms::Deploy)

The deploy API will continue to return the parent (deploy) task.  Using it,
the user may then request the child/sub-tasks using an API similar to the
following:

curl -k -u admin:password -H "Content-Type:application/json" -H "Accept:application/json,version=2" -X GET -d "{\"search\": \"parent_task_id=50b6b3ba-fa71-4a1c-840b-e1913fc5d8a6\"}" https://sat61fusor.example.com/foreman_tasks/api/tasks | json_reformat

Note: The above API will only return sub-actions that have begun processing.
This means that if ManageContent is executing the Rhev & CloudForms sub-actions
will not be available in the response.  Once ManageContent completes, the next
action (Rhev::Deploy) will also be available in subsequent API calls.

In addition to the above, the ManageContent has been restructured to
allow for it to perform the content actions for all products in the deployment
on a single pass versus the previous approach, which performed the actions on
a per-product basis (e.g. rhev, cloudforms...etc).
